requirements.txt
fastapi==0.115.5
jinja2==3.1.4
pydantic==2.10.1
llama-cpp-python==0.3.2
typing-extensions==4.12.2
runpod~=1.7.5
torch~=2.5.1